requirements.txt
aiosqlite==0.18.0
alembic==1.10.3
anyio==3.6.2
beautifulsoup4==4.12.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
databases==0.7.0
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
isort==5.12.0
Mako==1.2.4
MarkupSafe==2.1.2
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.47
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
